moved side classes to their own files
