+++ Editor/Database/GoogleAppScript/Util.cs.meta
@@ -1,3 +1,11 @@
-﻿fileFormatVersion: 2
-guid: b7557606df4e4a64a39cacf63fad5d7d
-timeCreated: 1619277451
+fileFormatVersion: 2
+guid: 4617fe2c39f6dff40a085ea4abc504c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
